# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "AddJobTextRequest": {
        "properties": {
          "company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company"
          },
          "jdText": {
            "title": "Jdtext",
            "type": "string"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "required": [
          "jdText"
        ],
        "title": "AddJobTextRequest",
        "type": "object"
      },
...
      "ApplicationUpsert": {
        "properties": {
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "ApplicationUpsert",
        "type": "object"
      },
...
      "BulkRequest": {
        "properties": {
          "action": {
            "enum": [
              "archive",
              "restore",
              "delete",
              "approve",
              "setStatus"
            ],
            "title": "Action",
            "type": "string"
          },
          "archived": {
            "default": false,
            "title": "Archived",
            "type": "boolean"
          },
          "board": {
            "enum": [
              "shortlist",
              "triage",
              "pipeline"
            ],
            "title": "Board",
            "type": "string"
          },
          "city": {
            "items": {
              "type": "string"
            },
            "title": "City",
            "type": "array"
          },
          "companySize": {
            "items": {
              "type": "string"
            },
            "title": "Companysize",
            "type": "array"
          },
          "country": {
            "items": {
              "type": "string"
            },
            "title": "Country",
            "type": "array"
          },
          "dryRun": {
            "default": true,
            "title": "Dryrun",
            "type": "boolean"
          },
          "employmentType": {
            "items": {
              "type": "string"
            },
            "title": "Employmenttype",
            "type": "array"
          },
          "ids": {
            "items": {
              "type": "integer"
            },
            "title": "Ids",
            "type": "array"
          },
          "industry": {
            "items": {
              "type": "string"
            },
            "title": "Industry",
            "type": "array"
          },
          "maxFit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Maxfit"
          },
          "minFit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Minfit"
          },
          "minSalary": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Minsalary"
          },
          "preset": {
            "default": "balanced",
            "title": "Preset",
            "type": "string"
          },
          "q": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Q"
          },
          "region": {
            "items": {
              "type": "string"
            },
            "title": "Region",
            "type": "array"
          },
          "remote": {
            "items": {
              "type": "string"
            },
            "title": "Remote",
            "type": "array"
          },
          "scope": {
            "enum": [
              "ids",
              "query"
            ],
            "title": "Scope",
            "type": "string"
          },
          "seniority": {
            "items": {
              "type": "string"
            },
            "title": "Seniority",
            "type": "array"
          },
          "skills": {
            "items": {
              "type": "string"
            },
            "title": "Skills",
            "type": "array"
          },
          "sortBy": {
            "default": "fit",
            "title": "Sortby",
            "type": "string"
          },
          "source": {
            "items": {
              "type": "string"
            },
            "title": "Source",
            "type": "array"
          },
          "sponsorship": {
            "items": {
              "type": "string"
            },
            "title": "Sponsorship",
            "type": "array"
          },
          "staleDays": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Staledays"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "statusIn": {
            "items": {
              "type": "string"
            },
            "title": "Statusin",
            "type": "array"
          }
        },
        "required": [
          "board",
          "action",
          "scope"
        ],
        "title": "BulkRequest",
        "type": "object"
      },
      "BulkResultOut": {
        "properties": {
          "affected": {
            "title": "Affected",
            "type": "integer"
          },
          "reasons": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Reasons",
            "type": "object"
          },
          "skipped": {
            "title": "Skipped",
            "type": "integer"
          }
        },
        "required": [
          "affected",
          "skipped",
          "reasons"
        ],
        "title": "BulkResultOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobDetail": {
        "properties": {
          "application": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApplicationOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "archivedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archivedat"
          },
          "bestResumeVersionId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bestresumeversionid"
          },
          "company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company"
          },
          "companySize": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Companysize"
          },
          "coverLetters": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/CoverLetterOut"
            },
            "title": "Coverletters",
            "type": "array"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "criteriaJson": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Criteriajson"
          },
          "employmentType": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Employmenttype"
          },
          "fitRationale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fitrationale"
          },
          "fitScore": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fitscore"
          },
          "hasProgress": {
            "title": "Hasprogress",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "industry": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Industry"
          },
          "jdText": {
            "title": "Jdtext",
            "type": "string"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "locationCity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locationcity"
          },
          "locationCountry": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locationcountry"
          },
          "locationRegion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locationregion"
          },
          "needsAttention": {
            "default": false,
            "title": "Needsattention",
            "type": "boolean"
          },
          "postedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Postedat"
          },
          "regressed": {
            "default": false,
            "title": "Regressed",
            "type": "boolean"
          },
          "rejectReason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rejectreason"
          },
          "remotePolicy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remotepolicy"
          },
          "resumeVersions": {
            "items": {
              "$ref": "#/components/schemas/ResumeVersionOut"
            },
            "title": "Resumeversions",
            "type": "array"
          },
          "salaryCurrency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Salarycurrency"
          },
          "salaryMax": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Salarymax"
          },
          "salaryMin": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Salarymin"
          },
          "seniority": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seniority"
          },
          "skills": {
            "items": {
              "$ref": "#/components/schemas/SkillTagOut"
            },
            "title": "Skills",
            "type": "array"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "sponsorshipSignal": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sponsorshipsignal"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "required": [
          "id",
          "source",
          "url",
          "company",
          "title",
          "location",
          "jdText",
          "status",
          "fitScore",
          "fitRationale",
          "criteriaJson",
          "postedAt",
          "archivedAt",
          "createdAt",
          "hasProgress",
          "application",
          "resumeVersions",
          "skills"
        ],
        "title": "JobDetail",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Resume Agent API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/jobs": {
      "post": {
        "operationId": "create_manual_job_api_jobs_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddJobTextRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Manual Job"
      }
    },
    "/api/jobs/bulk": {
      "post": {
        "operationId": "bulk_jobs_api_jobs_bulk_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkResultOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Bulk Jobs"
      }
    },
...
    "/api/jobs/{job_id}": {
      "delete": {
        "operationId": "delete_job_endpoint_api_jobs__job_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Job Endpoint"
      },
      "get": {
        "operationId": "get_job_detail_api_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job Detail"
      },
      "patch": {
        "operationId": "patch_job_api_jobs__job_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Job"
      }
    },
    "/api/jobs/{job_id}/application": {
      "put": {
        "operationId": "upsert_application_api_jobs__job_id__application_put",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationUpsert"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upsert Application"
      }
    },
...
    "/api/match-gap": {
      "get": {
        "operationId": "get_match_gap_api_match_gap_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchGapOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Match Gap"
      }
    },
    "/api/match-gap/refresh-clusters": {
      "post": {
        "operationId": "refresh_match_gap_clusters_api_match_gap_refresh_clusters_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh Match Gap Clusters"
      }
    },
...
  }
}